feat(applications): migrate to hybrid server/client architecture
Convert applications from full client component to Next.js hybrid pattern
following official migration guidelines for better performance and SEO.

- Extract DashboardContent to separate client component with SWR
- Convert main page to server component with static header shell
- Add Suspense boundary for progressive loading and streaming
- Implement proper SEO metadata for search engine optimization
- Maintain existing functionality while reducing client bundle size

The page now renders static elements immediately on the server while
streaming in dynamic content, providing better perceived performance
and actual Next.js benefits over a pure SPA approach.
